requirements.txt
pymupdf
camelot-py[cv]
pytesseract
pillow
pdf2image